Move Path internal class to parent
It'll need to be re-used for the PathOrderOptimizer as well once we merge all this to Arachne.

Contributes to issue CURA-7928.
